requirements.txt
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
firebase-admin
requests
